added needammo and alreadybuilt build sprites made buildables not get removed from available selection removed auto switch to another weapon any time buildable is selected, now should only auto switch once after the build finishes changed traceline mask for build info to mask_shot, so the sentry will get flipped if you try building up against non solids(like the windows in ff_dm, whatever they are)
notes: detpack is still not selectable if you dont have the ammo, do we want to change this like the sentry and dispenser? detpack is sorta like a use and lose weapon, unlike the slots for dispenser and sg, which are basically abilities that are slotted.

alreadybuilt/needammo sprites show up differently scaled between the sentry and dispenser, because the code draws the sprite at the location where the ghost would be. should we alter this in some way?