Suppressing .ain files from being generated: (See http://www.madabouthats.org/code-tf2/viewtopic.php?t=3664)
I commented out initialization calls to Source AI in world.cpp, and changed FAllowNPCs() to always return false (instead of true).  This doesn't appear to cause any problems on a listen server, but I guess we'll have to wait until this gets put on the Beta server to see if it helps (or hurts).