0001026: backpacks & flags can be discarded through doors
When thrown, I changed the flag's bounding box from being min(0,0,0) max(0,0,1) to the original dimensions, but with a height (z) value of 1.  In this way, flags can't get stuck in most entities (doors etc.) but it may require tweaking to make it fit through other gaps.

0001624: Flags, when dropped, do not respond to moving entities etc.

Removed the MOVETYPE_NONE call from CFFInfoScript::ResolveFlyCollisionCustom() so that the flags etc. correctly move with moving entities if dropped on a moving entity.