0001568: Falling into pitt damage shows electrocution icon, instead of falling damage icon
Added extra field to player_death event firing.  This field contains the damage type if it's a trigger_hurt causing the damage.  In the death notice function I then check if the killer is "trigger_hurt".  If so, I use the damagetype to match it to the appropriate death icon name.  Update your basemod too or you won't have the correct files.