0001683: Can use engineer radial menu when dead
Added a "IsAlive" check to the code that calls the menu in ff_player_shared.cpp. If the player is not alive, the menu display code is not called.

Not sure if this is the most correct way to do it, but it seems to have stopped the problem so... yeah.