Shield can now be more appropriately centered. ConVars include: ffdev_shield_build_center - 1/2 of the width ffdev_shield_build_distance - how far away from the player to build ffdev_shield_build_height - height off the ground
Shield can be toggled to be perpendicular with floor(like other ghost buildables), or use the players eye angles.  Im not sure which will be better.
Convars include:
ffdev_shield_vertical_motion 

Changed the RecalculateSpeed function in ff_player to allow the shield to slow to less than 40%.  It was previously hardcoded to not go under that unless u were shooting an AC.  Anybody know why?