Pipes now have 2 different detonation times depending on if they are currently magnetizing to a target or not.
ConVars include:
ffdev_pipebomb_detdelay - time for normal pipes to detonate ( defaulted to instant )
ffdev_pipebomb_magnetic_detdelay - time for magnetic pipes to detonate( defaulted .5s)

Pipes can not be detonated until they collide with something.  This is to prevent the unintuitive det delay confusion, and also so people can't insta-det pipes in the barrel.

Fixed a bug so pipes are forced to detonate when the player dies.