Fix silent grenades when client doesn't think you can prime yet
Sometimes it was possible to prime a grenade but not get any timers on the client, since the command would be sent to the server even though the client thinks the command is impossible. Example: If you immediately try to prime a gren as you spawn and your latency is high enough, then the client will think you're still dead and not start the timers, but by the time the command gets to the server it will think you're alive and prime it anyway

This fixes it so that if the client thinks you can't run a command, then the command is not sent to the server. Also updated all the shared commands with this fix so this might have some side-benefits to other things like cloaking.