Make extra `onbuild` context backwards compatible
Instead of changing the type of the parameters, the extra context is passed as a new second argument to the `onbuild` callback. The new second argument is a table with the keys `type` that contains a `BuildableTypes` constant and `fusetime` that contains a number with the fusetime of the detpack that is trying to be built (or nil if it's not a detpack being built).

Doing it this way also makes it easy to add more context in the future in a backwards compatible way (just add more stuff to the table).

Example usage:

```
function nobuild:onbuild( player, buildable_info )
  -- only allow detpacks if they have a fuse time of exactly 20
  if buildable_info.type == BuildableTypes.kDetpack and buildable_info.fusetime == 20 then
    return EVENT_ALLOWED
  end
  return EVENT_DISALLOWED
end
```